Create a test for createAdviser function and organize the project
