+++ src/characters-factory-functions.ts
@@ -1,32 +1,5 @@
-import { DexterityRange, Fighter } from "./characters/fighter/type";
-import type {
-  Adviser,
-  Character,
-  mandatoryData,
-  FidelityRange,
-  Squire,
-} from "./types";
-
-export const createAdviser = (
-  { name, surname, age }: mandatoryData,
-  advisee: Character
-): Adviser => {
-  const adviser: Adviser = {
-    name,
-    surname,
-    age,
-    isAlive: true,
-    advisee,
-    talk() {
-      return "No sé por qué, pero creo que voy a morir pronto";
-    },
-    die() {
-      adviser.isAlive = false;
-    },
-  };
-
-  return adviser;
-};
+import { Fighter } from "./characters/fighter/type";
+import type { mandatoryData, FidelityRange, Squire } from "./types";
 
 export const createSquire = (
   { name, surname, age }: mandatoryData,
