+++ src/types.ts
@@ -14,11 +14,6 @@ export interface Character extends mandatoryData {
   talk: () => string;
 }
 
-export interface Adviser extends Character {
-  advisee: Character;
-  talk: () => "No sé por qué, pero creo que voy a morir pronto";
-}
-
 export interface Squire extends Character {
   servedCharacter: Fighter;
   fidelity: FidelityRange;
